* UX improvement tracks page

* track uodate

* Track Page UX improvement

* popover added

* move to using socket instead of POST


* show build status on the frontend


* fix for non-file method and add screencast


add screencast to README


* also handle the zip download in the socket workflow


* multer not needed anymore


* add local jquery


* Smart footer

* small fixes

* Bug fix track

* hound issues

* var repeatation removed

* Add Underline on hover

* SocialIcons added

* speaker page updated

* popover.js  added

* Hound ci issues solved

* added tracks info

* Zip fix

* Extrafiles deleted

* FIXES

* Travis build check

* Audio as Zip

* added tracks title responsiveness

* Solved frontend issues

* Changes Frontend issues

* link home to index.html (sessions.html will be separate later)


* speakers page null image

* Card size made equal

* added placeholders to all breaked images

* added margin responsiveness

* Scroll removed

* Sorted dates

* fix

* Console Removed

* Add favicon (#531)


* Mobilefixes (#535)

* Zip fix

* Extrafiles deleted

* FIXES

* Travis build check

* Audio as Zip

* Iphone fixes

* Mention Errors on the Front End (#536)

Mention error explicitly on the front-end rather silently logging them
on server side

* Sponsor Images in Zip (#537)

* Sponsor fixed

* Fixed Sponsors

* Travis build

* delete unrequired main.js file (#539)


* Speaker page fixes (#545)

* Speaker fixes overflow

* extra code

* fix the json folder path in zip upload (#540)


* open live preview and download zip from new tab (#541)


* Added Event page  (#538)

* Added Event page

* Added carousel image

* Updated headings

* extra code removed

* Links in navbar corrected

* added speaker photos

* Speaker page

* Sponsor Section Added   (#554)

* Added Event page

* Added carousel image

* Updated headings

* extra code removed

* Links in navbar corrected

* added speaker photos

* Speaker page

* sponsors added

* Added SCSS

* send email after website is generated (#553)

+++ src/generator/backend/dist.js
@@ -57,6 +57,8 @@ module.exports = {
   },
   cleanDist: function(appFolder, err) {
     fs.emptyDir(distPath + '/' + appFolder, (emptyErr) => {
+      if(emptyErr)
+         return err(emptyErr);
       fs.remove(distPath + '/' + appFolder, err);
     });
   },
@@ -65,7 +67,8 @@ module.exports = {
     fs.mkdirpSync(distPath);
     fs.mkdirpSync(appPath);
     fs.mkdirpSync(appPath + '/audio');
-    fs.mkdirpSync(appPath + '/img/speakers');
+    fs.mkdirpSync(appPath + '/images/speakers');
+    fs.mkdirpSync(appPath + '/images/sponsors');
   },
   copyAssets: function(appFolder, err) {
     const appPath = distPath + '/' + appFolder;
@@ -75,7 +78,44 @@ module.exports = {
     const appPath = distPath + '/' + appFolder;
     fs.mkdirpSync(appPath + '/json');
     var zip = new admZip(uploadedFile);
-    zip.extractAllTo(appPath + '/json', true);
+     var zipEntries = zip.getEntries(); 
+
+     zipEntries.forEach(function(zipEntry) {
+     
+      switch(zipEntry.entryName){
+        case 'images/speakers/':
+        zip.extractEntryTo("images/speakers/", appPath ); 
+        break;
+        case 'images/sponsors/':
+        zip.extractEntryTo("images/sponsors/", appPath ); 
+        break;
+        case 'audio/':
+        zip.extractEntryTo("audio/", appPath);
+        break;
+        case 'sessions':
+        zip.extractEntryTo("sessions", appPath +'/json/');
+        break;
+        case 'speakers':
+        zip.extractEntryTo("speakers", appPath +'/json/');
+        break;
+        case 'microlocations' :
+        zip.extractEntryTo("microlocations", appPath+'/json/');
+        break;
+        case 'event' :
+        zip.extractEntryTo("event", appPath +'/json/');
+        break;
+        case 'sponsors' :
+        zip.extractEntryTo("sponsors", appPath +'/json/');
+        break;
+        case 'tracks':
+        zip.extractEntryTo("tracks", appPath +'/json/');
+        break;
+        default:
+      }
+  
+      
+    });
+    
   },
   fetchApiJsons: function(appFolder, apiEndpoint, done) {
     const endpoint = apiEndpoint.replace(/\/$/, '');
@@ -125,7 +165,16 @@ module.exports = {
   downloadSpeakerPhoto: function(appFolder, photoUrl) {
     const appPath = distPath + '/' +appFolder;
     const photoFileName = photoUrl.split('/').pop();
-    const photoFilePath = 'img/speakers/' + photoFileName;
+    const photoFilePath = 'images/speakers/' + photoFileName;
+
+    console.log('Downloading photo : ' + photoFileName);
+    downloadFile(photoUrl, appPath + '/' + photoFilePath);
+    return photoFilePath;
+  },
+   downloadSponsorPhoto: function(appFolder, photoUrl) {
+    const appPath = distPath + '/' +appFolder;
+    const photoFileName = photoUrl.split('/').pop();
+    const photoFilePath = 'images/sponsors/' + photoFileName;
 
     console.log('Downloading photo : ' + photoFileName);
     downloadFile(photoUrl, appPath + '/' + photoFilePath);
